ffun/pyproject.toml
[tool.poetry]
name = "ffun"
version = "0.4.0"
description = "Backend for the Feeds Fun — web-based news reader"
repository = "https://github.com/Tiendil/feeds.fun"
authors = ["Aliaksei Yaletski (Tiendil) <a.eletsky@gmail.com>"]
license = " BSD-3-Clause"
readme = "README.md"
homepage = "https://feeds.fun"
keywords = ["news", "news-reader", "news-aggregator",
            "rss", "rss-reader", "rss-aggregator",
            "feed", "feed-reader", "feed-aggregator",
            "atom", "self-hosted"]
classifiers = [
        "Environment :: Web Environment",
        "Framework :: AsyncIO",
        "Framework :: FastAPI",
        "Intended Audience :: End Users/Desktop",
        "Intended Audience :: Developers",
        "License :: OSI Approved :: BSD License",
        "Natural Language :: English",
        "Operating System :: OS Independent",
        "Programming Language :: Python :: 3.11",
        "Programming Language :: Python :: 3 :: Only",
        "Programming Language :: Python :: Implementation :: CPython",
        "Topic :: Communications",
        "Topic :: Office/Business :: News/Diary",
        "Topic :: Text Processing",
        "Topic :: Text Processing :: Linguistic",
        "Topic :: Internet :: WWW/HTTP",
        "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
        "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
        "Topic :: Internet :: WWW/HTTP :: WSGI :: Server",
        "Topic :: Internet :: WWW/HTTP :: HTTP Servers"]


[tool.poetry.dependencies]
python = "^3.11"

uvicorn = {extras = ["standard"], version = "0.*"}
fastapi = "0.100.*"
pydantic = "2.1.*"
pydantic-settings = "2.0.*"
typer = "0.9.*"
httpx = "0.24.*"

openai = "0.27.*"
tiktoken = "0.4.*"

bidict = "0.22.*"

feedparser = "6.0.*"

beautifulsoup4 = "4.12.*"

async-lru = "2.0.*"

psycopg = "3.1.*"
psycopg-pool = "3.1.*"
yoyo-migrations = "8.2.*"

structlog = "23.1.*"

yarl = "1.9.*"

python-multipart = "0.0.*"

supertokens-python = "0.14.*"

python-slugify = "8.0.*"

sentry-sdk = "1.29.*"

tabulate = "0.9.*"

cryptography = "41.0.*"

markdown = "3.4.*"

[tool.poetry.dev-dependencies]

pytest = "7.4.*"
pytest-asyncio = "0.21.*"
pytest-mock = "3.11.*"

isort = "5.12.*"
black = "23.7.*"

flake8 = "6.1.*"
flake8-docstrings = "1.7.*"
flake8-functions = "0.0.*"
flake8-annotations-complexity = "0.0.*"
flake8-cognitive-complexity = "0.1.*"
flake8-pytest = "1.4.*"
flake8-bandit = "4.1.*"
flake8-absolute-import = "1.0.*"
flake8-print = "5.0.*"
flake8-pyproject = "1.2.*"
flake8-eradicate = "1.5.*"
autoflake = "2.2.*"

# mypy fails in the latest versions
mypy = "1.5.0"

types-python-slugify = "8.0.*"
types-beautifulsoup4 = "4.12.*"
types-markdown = "3.4.*"
types-pkg-resources = "0.1.*"
types-tabulate = "0.9.*"

[tool.poetry.scripts]

ffun = "ffun.cli.cli:app"

[tool.isort]
profile = "black"
line_length = 119

[tool.black]
line-length = 119

[tool.mypy]
plugins = ["pydantic.mypy"]
strict = true
implicit_reexport = true
allow_untyped_calls = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[[tool.mypy.overrides]]
module = "yoyo.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "feedparser.*"
ignore_missing_imports = true

# remove after updating to a new version
[[tool.mypy.overrides]]
module = "supertokens_python.*"
ignore_errors = true

[tool.pytest.ini_options]
addopts = """-x -vv --strict-markers -p no:cacheprovider"""
filterwarnings = [
   "error",
    "ignore::DeprecationWarning:feedparser.encodings:29",
    "ignore::DeprecationWarning:yoyo.migrations:34",
]

[tool.flake8]
ignore = [
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D107",
    "D200",
    "D400",
    "S101",
    "S311",
    "W503",
    "CFQ004",
    "E203",  # blake formats [x:y] with spaves
]
max-line-length = 119


[tool.autoflake]
in_place = true
remove_all_unused_imports = true
remove_unused_variables = true
recursive = true
